Add .php_cs for code style and fix phpunit.xml for tests
Remove DB_CONNECTION and DB_DATABASE from the phpunit.xml so that is
possible to set it up with .env.testing file instead. This way the
developer can choose the database for testing environment to run the
tests in a real database engine instead of in memory sqlite.
This is a good practice and is recommended by Laravel community.

Also adds a helper file in app/Support folder to provide many helper
functions. The first function given is 'money' helper to convert a
number to currency format string. This was created because the current
money_format php function is deprecated in php 7.4.